Conflicts:
	Makefile
	src/game.c

+++ src/player.h
@@ -15,7 +15,7 @@
 typedef struct _Player {
 	SDL_Window * window;
 	SDL_Rect * rect;
-	Uint8 oxygen;
+	float oxygen;
 	SDL_Rect * sprite_rect;
 	SDL_Surface * surface;
 	Uint32 time_between_shots;
